Moved API fetching to App.vue and store them til pinia on response instead
