+++ frontend/src/stores/github.ts
@@ -1,7 +1,4 @@
 import { defineStore } from "pinia";
-import axios from 'axios';
-import { ref, getDownloadURL, listAll } from "firebase/storage";
-import { storage } from '@/firebase';
 
 export const useGitHubStore = defineStore('github', {
   state: () => ({
@@ -18,58 +15,10 @@ export const useGitHubStore = defineStore('github', {
     numberOfFollowers: (state) => state.followers.length,
   },
   actions: {
-    async fetchRepositories() {
-      try {
-        console.info('Fetching repositories from GitHub API')
-        const data = await axios.get('https://api.github.com/users/ipeglin/repos');
-        this.repositories = data.data;
-      } catch (error) {
-        console.error(error);
-      }
-    },
-    async fetchFollowers() {
-      try {
-        console.info('Fetching followers from GitHub API')
-        const data = await axios.get('https://api.github.com/users/ipeglin/followers');
-        this.followers = data.data;
-      } catch (error) {
-        console.error(error);
-      }
-    },
     async fetchRepository(repoName: string) {
-      try {
-        console.info('Fetching repository information from GitHub API')
-        const data = await axios.get(`https://api.github.com/repos/ipeglin/${repoName}`)
-        this.currentRepository = data.data;
-      } catch (error) {
-        console.error(error);
-      }
-    },
-    async fetchRepoImages() {
-      if (Object.keys(this.repoImages).length  !== 0) return console.info('Images have allready been fetched');
+      if (this.repositories.length === 0) return;
 
-      console.info('Fetching repo images from database')
-
-      // Create a reference to the file we want to download
-      // const storage = getStorage();
-      const listRef = ref(storage, 'images/repos');
-
-      const repoNames = await this.repos.map((repo: any) => repo.name);
-
-      listAll(listRef)
-        .then((res) => {
-          res.items.forEach((itemRef) => {
-            const imageName = itemRef.name.replace(/\.[^/.]+$/, "");
-            if (repoNames.includes(imageName)) {
-              getDownloadURL(itemRef)
-                .then((url) => this.repoImages[imageName] = url);
-            }
-          })
-          this.images;
-        })
-        .catch((error) => {
-          console.error(error);
-        })
-    },
+      return this.repositories.filter((repo: any) => repo.name = repoName);
+    }
   }
 })
